move timeline to a shared component
